requirements.txt
Flask~=2.0.3
click~=8.0.4
Werkzeug~=2.0.3
setuptools~=57.0.0
python-dotenv~=0.19.2